:tractor: relocated data loader manager to core folder
