+++ moban/core/context.py
@@ -4,7 +4,7 @@ import copy
 from moban import utils, constants, exceptions
 from moban.externals import reporter
 from moban.program_options import OPTIONS
-from moban.data_loaders.manager import merge, load_data
+from moban.core.data_loader import merge, load_data
 
 
 class Context(object):
